+++ app/system/modules/info/views/info.php
@@ -5,10 +5,10 @@
 
         <div class="uk-panel">
             <ul class="uk-nav uk-nav-default pk-nav-large" uk-switcher="connect: #tab-content">
-                <li class="uk-active"><a><span class="pk-icon-large-server uk-margin-right"></span><span class="uk-text-middle">{{ 'System' | trans }}</span></a></li>
-                <li><a><span class="pk-icon-large-code uk-margin-right"></span><span class="uk-text-middle">{{ 'PHP' | trans }}</span></a></li>
-                <li><a><span class="pk-icon-large-database uk-margin-right"></span><span class="uk-text-middle">{{ 'Database' | trans }}</span></a></li>
-                <li><a><span class="pk-icon-large-lock-file uk-margin-right"></span><span class="uk-text-middle">{{ 'Permissions' | trans }}</span></a></li>
+                <li><a><span class="uk-margin-right" uk-icon="icon:router;ratio:1.3"></span><span class="uk-text-middle">{{ 'System' | trans }}</span></a></li>
+                <li><a><span class="uk-margin-right" uk-icon="icon:php;ratio:1.3"></span><span class="uk-text-middle">{{ 'PHP' | trans }}</span></a></li>
+                <li><a><span class="uk-margin-right" uk-icon="icon:database;ratio:1.3"></span><span class="uk-text-middle">{{ 'Database' | trans }}</span></a></li>
+                <li><a><span class="uk-margin-right" uk-icon="icon:face-id;ratio:1.3"></span><span class="uk-text-middle">{{ 'Permissions' | trans }}</span></a></li>
             </ul>
         </div>