+++ app/system/modules/settings/views/settings.php
@@ -8,7 +8,11 @@
             <div class="uk-panel">
 
                 <ul class="uk-nav uk-nav-default pk-nav-large" ref="tab">
-                    <li v-for="section in orderBy(sections, 'priority')" :key="section.name"><a><span :class="'uk-margin-right ' + section.icon"></span><span class="uk-text-middle">{{ section.label | trans }}</span></a></li>
+                    <li v-for="section in orderBy(sections, 'priority')" :key="section.name">
+                        <a>
+                            <i class="uk-margin-right" :uk-icon="`icon:${section.icon};ratio:1.3`"></i><span class="uk-text-middle">{{ section.label | trans }}</span>
+                        </a>
+                    </li>
                 </ul>
 
             </div>